pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.2.0"
description = "lookup-poetry"
authors = ["weverhall <weverhall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.7.0"
validators = "^0.18.2"
python-whois = "^0.7.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
